+++ rollup.common.js
@@ -35,7 +35,7 @@ export default ({ isDev = false } = {}) => {
 
 	return {
 		input: 'src/index.ts',
-		external: ['react', 'css/debug.css'],
+		external: ['react'],
 		output,
 		plugins: [
 			progress(),
@@ -45,6 +45,7 @@ export default ({ isDev = false } = {}) => {
 			commonjs({ sourceMap, extensions }),
 			postcss({
 				inject: true,
+				modules: true,
 				sourceMap,
 				exclude: 'node_modules/**',
 				plugins: [pcssPresetEnv(), autoprefixer(), !isDev && cssnano()],